requirements.txt
Flask
SQLAlchemy
psycopg2
langchain
langchain-community
langchain-huggingface
langchain-cohere
langchain-openai
langchain-anthropic
langchain-mistralai
langchain-google-genai
langchain-deepseek
langgraph
huggingface-hub
sentence-transformers
faiss-cpu
cohere
soundfile
numpy
uuid
evaluate
ragas
datasets
torch
transformers
python-dotenv
matplotlib
seaborn
